add spring security

+++ src/main/java/com/example/entities/User.java
@@ -1,5 +1,6 @@
 package com.example.entities;
 
+
 import javax.persistence.*;
 
 @Entity
@@ -7,7 +8,7 @@ import javax.persistence.*;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
+    private int id;
 
     @Column(nullable = false, unique = true)
     private String userName;
@@ -15,11 +16,11 @@ public class User {
     @Column(nullable = false)
     private String password;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
